# Avoid parsing numbers when using RPCv2 protocol